Manny DeMaio
view req session panel now updates the fields correctly when clicking on the requirement name.

Save requirment button only works on the all req panel.  The session that is sent is session 1,
where all the requirments are held, making it not be able to search the current sessions requirements and update them.